Adedd insert creation time in converted videos (ffmpeg param)

+++ Shotwell_event2folder.py
@@ -771,26 +771,39 @@ if __name__ == '__main__':
 					Entry_title = entry [15]
 					Entry_comment = entry [19]
 
-					Entry_tag_id = 'video-%016x,'%Entry_id
-
 					if itemcheck (sourcefile) != 'file':
 						logging.warning ('\tThis file cannot be accessed, or does not exist at this very moment.')
 						continue
-					
+										
+					Entry_tag_id = 'video-%016x,'%Entry_id
+					metadataparam = ''
+					if Entry_exposure_time != 0:
+						videoCreationTime = datetime.fromtimestamp (Entry_exposure_time)
+						videoStringTime = datetime.isoformat(videoCreationTime, timespec='microseconds') + 'Z'  # Example:   2018-01-03T18:25:34.000000Z
+						metadataparam += '-metadata creation_time="{}"'.format(videoStringTime)
+
 					logging.info ('Processing file with ffmpeg: {}'.format(sourcefile))
 					newFilename = os.path.splitext(sourcefile)[0]+'_c.mov'
 					if itemcheck (newFilename) == 'file':
-						os.remove(newFilename)
-						logging.warning ('\tIt seems that an old converted file was there, it has been deleted.')
+						if dummy == False:
+							os.remove(newFilename)
+						logging.warning ('\tIt seems that an old converted file was there, it has been deleted.{}'.format(dummymsg))
 					
-					ffmpeg_status = os.system ('ffmpeg -i "{}" "{}"'.format(sourcefile,newFilename))
-					logging.debug ('\tffmpeg exitted with code: {}'.format(ffmpeg_status))
+					if dummy == False:
+						ffmpeg_status = os.system ('ffmpeg -i "{}" {} "{}"'.format(sourcefile, metadataparam, newFilename))
+					else:
+						ffmpeg_status = 0
+					logging.debug ('\tffmpeg exitted with code: {}{}'.format(ffmpeg_status, dummymsg))
 					if ffmpeg_status == 0:
 						# (ffmpeg exitted with no errors)
 						logging.info ('\tFile converted, adding or updating new entries to DB')
 						# Getting new values for update DB registry.
-						newMD5 = md5hash (newFilename)
-						newFilesize = os.path.getsize (newFilename)
+						if dummy == False:
+							newMD5 = md5hash (newFilename)
+							newFilesize = os.path.getsize (newFilename)
+						else:
+							newMD5 = 0
+							newFilesize = 0
 						newEntry = (None,
 									newFilename,
 									Entry_width,
@@ -813,38 +826,42 @@ if __name__ == '__main__':
 									)
 						# Fetching videoentry for an already converted video.
 						videoConvlineID = dbconnection.execute ('SELECT id FROM videotable WHERE filename=? ', (newFilename,)).fetchone()
-						print ('videoConvlineID:', videoConvlineID)
 						if videoConvlineID is None:
-							logging.debug ('Inserting new line at VideoTable')
-							dbconnection.execute ('INSERT INTO videotable VALUES (?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?) ', newEntry )
-							# Adding new videofiles to tag table (cloning values)
-							dbconnection.commit ()
-							newEntry_id = dbconnection.execute ('SELECT max(id) FROM videotable').fetchone()[0]
-							newVideoTag_id = 'video-%016x'%newEntry_id + ','
-							TagCursor = dbconnection.cursor ()
-							TagCursor.execute ("SELECT id, photo_id_list FROM tagtable WHERE photo_id_list LIKE '%{}%'".format(Entry_tag_id,))
-							for TagEntry in TagCursor:
-								lineID , tagtext = TagEntry[0], TagEntry[1]
-								newTagText = tagtext + newVideoTag_id
-								dbconnection.execute ('UPDATE tagtable SET photo_id_list=? WHERE id=?',(newTagText,lineID))
+							logging.debug ('Inserting new line at VideoTable.{}'.format(dummymsg))
+							if dummy == False:
+								dbconnection.execute ('INSERT INTO videotable VALUES (?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?) ', newEntry )
+								# Adding new videofiles to tag table (cloning values)
+								dbconnection.commit ()
+								newEntry_id = dbconnection.execute ('SELECT max(id) FROM videotable').fetchone()[0]
+								newVideoTag_id = 'video-%016x'%newEntry_id + ','
+								TagCursor = dbconnection.cursor ()
+								TagCursor.execute ("SELECT id, photo_id_list FROM tagtable WHERE photo_id_list LIKE '%{}%'".format(Entry_tag_id,))
+								for TagEntry in TagCursor:
+									lineID , tagtext = TagEntry[0], TagEntry[1]
+									newTagText = tagtext + newVideoTag_id
+									dbconnection.execute ('UPDATE tagtable SET photo_id_list=? WHERE id=?',(newTagText,lineID))
 
 						else:
-							logging.debug ('Updating an existent registry for converted video.')
+							logging.debug ('Updating an existent registry for converted video.{}'.format(dummymsg))
 							# This will not update or clone the tag registry, it will preserve existent converted video tag attributes and rating.
-							dbconnection.execute ('UPDATE videotable SET filesize=?, import_id=?, md5=?, time_created=? WHERE id = ?', (newFilesize, newImportID, newMD5, int(now.timestamp()), videoConvlineID[0]))
+							if dummy == False:
+								dbconnection.execute ('UPDATE videotable SET filesize=?, import_id=?, md5=?, time_created=? WHERE id = ?', (newFilesize, newImportID, newMD5, int(now.timestamp()), videoConvlineID[0]))
 						
 						# Set original video as rejected. (rating = -1)
-						dbconnection.execute ('UPDATE videotable SET rating=-1 WHERE id = ?', (Entry_id,))
+						if dummy == False:
+							dbconnection.execute ('UPDATE videotable SET rating=-1 WHERE id = ?', (Entry_id,))
 						
 					else:
 						# ffmpeg encounterered errors
-						if itemcheck (newFilename) == 'file':
-							os.remove(newFilename)
-						failedName = os.path.splitext(sourcefile)[0]+'_f.mov'
-						os.rename (sourcefile, failedName)
-						dbconnection.execute('UPDATE videotable SET filename=? WHERE id=?', (failedName,Entry_id))
+						if dummy == False:
+							if itemcheck (newFilename) == 'file':
+								os.remove(newFilename)
+							failedName = os.path.splitext(sourcefile)[0]+'_f.mov'
+							os.rename (sourcefile, failedName)
+							dbconnection.execute('UPDATE videotable SET filename=? WHERE id=?', (failedName,Entry_id))
 
-					dbconnection.commit()
+					if dummy == False:
+						dbconnection.commit()
 				dbMOVcursor.close()
 			# Closing db Connection
 			dbconnection.close ()
